Histogram

+++ src/test/java/Funkcje/Srednia.java
@@ -2,8 +2,8 @@ package Funkcje;
 
 public class Srednia {
 
-	public double wyliczona_srednia(int[] dane) {
-		int suma = 0;
+	public double wyliczona_srednia(double [] dane) {
+		double suma = 0;
 		double srednia = 0;
 		int rozmiar = dane.length;
 		for (int i = 0; i < rozmiar; i++) {
